# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basirah API Documentation",
        "description": "API documentation for Basirah backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/admin/login": {
            "post": {
                "tags": [
                    "AdminAuth"
                ],
                "summary": "Admin login",
                "operationId": "6e9fa1e6623769561167fc73fb39451e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/admin/profile": {
            "get": {
                "tags": [
                    "AdminAuth"
                ],
                "summary": "Get admin profile",
                "operationId": "be76cdbb973229d91dac2c9903626323",
                "responses": {
                    "200": {
                        "description": "Admin profile returned"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "AdminAuth"
                ],
                "summary": "Update admin profile (Super Admin only)",
                "operationId": "bd90a50ca8b594f804fd7fdc4e7df96e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    },
                    "403": {
                        "description": "Only super admins can edit their profile."
                    }
                }
            }
        },
        "/admin/change-password": {
            "post": {
                "tags": [
                    "AdminAuth"
                ],
                "summary": "Change admin password",
                "operationId": "1b1cc7e810c1adb5040ae3a2afefa2fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "400": {
                        "description": "Current password is incorrect"
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "AdminAuth"
                ],
                "summary": "Admin logout",
                "operationId": "be5acf48746a8be5203223ec84a10b89",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                }
            }
        },
        "/": {
            "get": {
                "summary": "API Root",
                "operationId": "113b4f9fb7de60f6084f5aed4ed1b140",
                "responses": {
                    "200": {
                        "description": "API is running"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "AdminAuth",
            "description": "Admin Authentication and Profile management"
        }
    ]
}